+++ generate_reflection/src/database.rs
@@ -1,11 +1,15 @@
-use std::collections::HashMap;
+use std::{
+    borrow::Cow,
+    collections::HashMap,
+};
 
-use rbx_dom_weak::RbxValue;
-
-use crate::api_dump::Dump;
+use crate::{
+    api_dump::Dump,
+    reflection_types::RbxInstanceClass,
+};
 
 pub struct ReflectionDatabase {
     pub dump: Dump,
-    pub default_properties: HashMap<String, HashMap<String, RbxValue>>,
     pub studio_version: [u32; 4],
+    pub classes: HashMap<Cow<'static, str>, RbxInstanceClass>,
 }