+++ generate_reflection/src/reflection_types.rs
@@ -1,17 +1,77 @@
+//! Wraps the reflection_types module from rbx_reflection and implements extra
+//! methods its types.
+
+use std::borrow::Cow;
+
+use log::warn;
 use rbx_dom_weak::RbxValueType;
 
 use crate::api_dump::{ValueType, ValueCategory};
 
-#[derive(Debug, Clone, PartialEq)]
-pub enum RbxPropertyType<'a> {
-    Data(RbxValueType),
-    Enum(&'a str),
+// The types this module exposes are defined in the rbx_reflection crate itself
+// since they're supposed to line up 1:1 and we don't want them to get out of
+// date.
+//
+// Normally, we'd want to use &'static str for strings in the generated code and
+// String in our generator. We compromise and use Cow<'static, str> instead!
+#[path = "../../rbx_reflection/src/reflection_types.rs"]
+mod inner;
 
-    UnimplementedType(&'a str),
+pub use inner::*;
+
+impl RbxInstanceTags {
+    pub fn from_dump_tags<T: AsRef<str>>(dump_tags: &[T]) -> RbxInstanceTags {
+        let mut tags = RbxInstanceTags::empty();
+
+        for dump_tag in dump_tags {
+            let converted = match dump_tag.as_ref() {
+                "Deprecated" => RbxInstanceTags::DEPRECATED,
+                "NotBrowsable" => RbxInstanceTags::NOT_BROWSABLE,
+                "NotCreatable" => RbxInstanceTags::NOT_CREATABLE,
+                "NotReplicated" => RbxInstanceTags::NOT_REPLICATED,
+                "PlayerReplicated" => RbxInstanceTags::PLAYER_REPLICATED,
+                "Service" => RbxInstanceTags::SERVICE,
+                "Settings" => RbxInstanceTags::SETTINGS,
+                _ => {
+                    warn!("Unknown instance flag {}", dump_tag.as_ref());
+                    continue;
+                }
+            };
+
+            tags |= converted;
+        }
+
+        tags
+    }
 }
 
-impl<'a> From<&'a ValueType> for RbxPropertyType<'a> {
-    fn from(value_type: &'a ValueType) -> RbxPropertyType<'a> {
+impl RbxPropertyTags {
+    pub fn from_dump_tags<T: AsRef<str>>(dump_tags: &[T]) -> RbxPropertyTags {
+        let mut tags = RbxPropertyTags::empty();
+
+        for dump_tag in dump_tags {
+            let converted = match dump_tag.as_ref() {
+                "Deprecated" =>  RbxPropertyTags::DEPRECATED,
+                "Hidden" =>  RbxPropertyTags::HIDDEN,
+                "NotBrowsable" =>  RbxPropertyTags::NOT_BROWSABLE,
+                "NotReplicated" =>  RbxPropertyTags::NOT_REPLICATED,
+                "NotScriptable" =>  RbxPropertyTags::NOT_SCRIPTABLE,
+                "ReadOnly" =>  RbxPropertyTags::READ_ONLY,
+                _ => {
+                    warn!("Unknown instance flag {}", dump_tag.as_ref());
+                    continue;
+                }
+            };
+
+            tags |= converted;
+        }
+
+        tags
+    }
+}
+
+impl<'a> From<&'a ValueType> for RbxPropertyType {
+    fn from(value_type: &'a ValueType) -> RbxPropertyType {
         match value_type.category {
             ValueCategory::Primitive => {
                 let data_kind = match value_type.name.as_str() {
@@ -19,15 +79,13 @@ impl<'a> From<&'a ValueType> for RbxPropertyType<'a> {
                     "string" => RbxValueType::String,
                     "int" => RbxValueType::Int32,
                     "float" => RbxValueType::Float32,
-
-                    // These aren't quite right:
-                    "double" => RbxValueType::Float32,
+                    "double" => RbxValueType::Float64,
                     "int64" => RbxValueType::Int32,
 
                     unknown => {
                         println!("Can't emit primitives of type {}", unknown);
 
-                        return RbxPropertyType::UnimplementedType(&value_type.name);
+                        return RbxPropertyType::UnimplementedType(Cow::Owned(value_type.name.to_owned()));
                     },
                 };
 
@@ -43,17 +101,27 @@ impl<'a> From<&'a ValueType> for RbxPropertyType<'a> {
                     "BinaryString" => RbxValueType::BinaryString,
                     "UDim" => RbxValueType::UDim,
                     "UDim2" => RbxValueType::UDim2,
+                    "Content" => RbxValueType::Content,
+                    "NumberRange" => RbxValueType::NumberRange,
+                    "NumberSequence" => RbxValueType::NumberSequence,
+                    "ColorSequence" => RbxValueType::ColorSequence,
+                    "ProtectedString" => RbxValueType::String,
+
+                    "QDir" | "QFont" => {
+                        // We're never going to support these types.
+                        return RbxPropertyType::UnimplementedType(Cow::Owned(value_type.name.to_owned()));
+                    },
 
                     unknown => {
                         println!("Can't emit data of type {}", unknown);
 
-                        return RbxPropertyType::UnimplementedType(&value_type.name);
+                        return RbxPropertyType::UnimplementedType(Cow::Owned(value_type.name.to_owned()));
                     },
                 };
 
                 RbxPropertyType::Data(data_kind)
             }
-            ValueCategory::Enum => RbxPropertyType::Enum(&value_type.name),
+            ValueCategory::Enum => RbxPropertyType::Enum(Cow::Owned(value_type.name.to_owned())),
             ValueCategory::Class => RbxPropertyType::Data(RbxValueType::Ref),
         }
     }