+++ rbx_reflection/src/lib.rs
@@ -1,12 +1,9 @@
-mod core;
 mod reflection_database;
+mod reflection_types;
 mod resolution;
-mod types;
-mod version;
 
 pub use crate::{
-    core::{get_classes, get_enums},
+    reflection_database::*,
+    reflection_types::*,
     resolution::{try_resolve_value, ValueResolveError},
-    types::*,
-    version::*,
 };